dse: Refactoring (add clusterizer, stencil calculation, ...)
