+++ devito/dse/__init__.py
@@ -1,4 +1,5 @@
 from devito.dse.aliases import *  # noqa
+from devito.dse.clusterizer import *  # noqa
 from devito.dse.extended_sympy import *  # noqa
 from devito.dse.graph import *  # noqa
 from devito.dse.inspection import *  # noqa
