+++ devito/stencilkernel.py
@@ -13,7 +13,8 @@ from devito.cgen_utils import Allocator, blankline
 from devito.compiler import (get_compiler_from_env, jit_compile, load)
 from devito.dimension import BufferedDimension, Dimension, time
 from devito.dle import compose_nodes, filter_iterations, transform
-from devito.dse import (estimate_cost, estimate_memory, indexify, rewrite)
+from devito.dse import (clusterize, estimate_cost, estimate_memory, indexify,
+                        rewrite, q_indexed)
 from devito.interfaces import SymbolicData, Forward, Backward
 from devito.logger import bar, error, info, info_at
 from devito.nodes import (Element, Expression, Function, Iteration, List,
@@ -70,18 +71,27 @@ class OperatorBasic(Function):
         # Set the direction of time acoording to the given TimeAxis
         time.reverse = time_axis == Backward
 
-        # Normalize the expressions
+        # Expression lowering
         expressions = [indexify(s) for s in as_tuple(expressions)]
         expressions = [s.xreplace(subs) for s in expressions]
 
-        # Retrieve the data type of the Operator
+        # Analysis 1 - required *also after* the StencilKernel construction
         self.dtype = self._retrieve_dtype(expressions)
+        self.output = self._retrieve_output_fields(expressions)
+
+        # Analysis 2 - required *for* the StencilKernel construction
+        ordering = self._retrieve_loop_ordering(expressions)
+        stencils = self._retrieve_stencils(expressions)
 
         # Apply the Devito Symbolic Engine for symbolic optimization
-        dse_state = rewrite(expressions, mode=dse)
+        expressions = rewrite(expressions, stencils, mode=dse)
+
+        # Group expressions based on their stencil
+        clusters = clusterize(expressions, stencils)
+        from IPython import embed; embed()
 
         # Wrap expressions with Iterations according to dimensions
-        nodes = self._schedule_expressions(dse_state)
+        nodes = self._schedule_expressions(clusters, ordering)
 
         # Introduce C-level profiling infrastructure
         self.sections = OrderedDict()
@@ -108,8 +118,7 @@ class OperatorBasic(Function):
         nodes, elemental_functions = self._insert_declarations(dle_state, parameters)
         self.elemental_functions = elemental_functions
 
-        # Track the DSE and DLE output, as they may be useful later
-        self._dse_state = dse_state
+        # Track the DLE output, as it might be useful at execution time
         self._dle_state = dle_state
 
         # Finish instantiation
@@ -284,22 +293,16 @@ class OperatorBasic(Function):
         as if blocking were not applied (ie, unitary block size)."""
         pass
 
-    def _schedule_expressions(self, dse_state):
-        """Wrap :class:`Expression` objects within suitable hierarchies of
-        :class:`Iteration` objects according to dimensions and stencils."""
+    def _schedule_expressions(self, clusters, ordering):
+        """Wrap :class:`Expression` objects, already grouped in :class:`Cluster`
+        objects, within nested :class:`Iteration` objects (representing loops),
+        according to dimensions and stencils."""
         # Will be used to filter out aliasing buffered dimensions
         key = lambda d: d.parent if d.is_Buffered else d
 
-        # Establish a partial ordering for the Iterations based on the order
-        # by which dimensions appeared in the input expressions
-        ordering = []
-        for i in flatten(Stencil(i).dimensions for i in dse_state.input):
-            if i not in ordering:
-                ordering.extend([i, i.parent] if i.is_Buffered else [i])
-
         processed = []
         schedule = OrderedDict()
-        for c in dse_state.clusters:
+        for i in clusters:
             # Build the Expression objects to be inserted within an Iteration tree
             expressions = [Expression(v, np.int32 if c.trace.is_index(k) else self.dtype)
                            for k, v in c.trace.items()]
@@ -399,6 +402,32 @@ class OperatorBasic(Function):
         if len(lhss) != 1:
             raise RuntimeError("Expression types mismatch.")
         return lhss.pop()
+
+    def _retrieve_loop_ordering(self, expressions):
+        """
+        Establish a partial ordering for the loops that will appear in the code
+        generated by the StencilKernel, based on the order in which dimensions
+        appear in the input expressions.
+        """
+        ordering = []
+        for i in flatten(Stencil(i).dimensions for i in expressions):
+            if i not in ordering:
+                ordering.extend([i, i.parent] if i.is_Buffered else [i])
+        return ordering
+
+    def _retrieve_output_fields(self, expressions):
+        """Retrieve the fields computed by the StencilKernel."""
+        return [i.lhs.base.function for i in expressions if q_indexed(i.lhs)]
+
+    def _retrieve_stencils(self, expressions):
+        """Determine the :class:`Stencil` of each provided expression."""
+        mapper = OrderedDict()
+        for i in expressions:
+            mapper.setdefault(i.lhs, []).append(i)
+        stencils = OrderedDict()
+        for k, v in mapper.items():
+            stencils[k] = Stencil.union(*[Stencil(i) for i in v])
+        return stencils
 
     @property
     def _cparameters(self):
@@ -533,7 +562,7 @@ class OperatorCore(OperatorBasic):
         at_arguments = arguments.copy()
 
         # Output data must not be changed
-        output = [i.name for i in self._dse_state.output_fields]
+        output = [i.name for i in self.output]
         for k, v in arguments.items():
             if k in output:
                 at_arguments[k] = v.copy()
